+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # PyProbables Changelog
 
+### Version 0.1.1:
+* CuckooFilter
+    * Import / Export functionality
+    * Enforce single insertion per key
+    * Auto expand when insertion failure OR when called to do so (settable)
+
 ### Version 0.1.0:
 * Cuckoo Filter
     * Added basic Cuckoo Filter code